backend/requirements.txt
Django==4.2.7
djangorestframework==3.15.2
django-cors-headers==4.8.0
requests==2.31.0
psutil==5.9.6
GPUtil==1.4.0
wmi==1.5.1
cryptography==41.0.7
python-dotenv>=1.0.0
google-generativeai>=0.3.0

# AutoGen Dependencies for MCP Task Automation
# Note: AutoGen agents mode is optional - direct execution works without these
# Uncomment if you want to use AutoGen agents instead of direct execution
# pyautogen==0.2.18
# openai==1.12.0

# Advanced Telemetry Dependencies (Optional - for HWiNFO-level sensor monitoring)
# Uncomment and install for deep hardware telemetry:
# pythonnet>=3.0.0              # For LibreHardwareMonitor integration
# nvidia-ml-py3>=7.352.0        # For NVIDIA GPU telemetry via NVML
